- Adjusted env - Changed some imports - Evaluation now happens with pytorch - FullEvalDataset returns label in float now - Adjusted optimaing metric

+++ data/dataset.py
@@ -201,4 +201,4 @@ class FullEvalDataset(data.Dataset):
         return self.n_users
 
     def __getitem__(self, user_index):
-        return user_index, np.arange(self.n_items), self.evaluation_matrix[user_index].toarray().squeeze()
+        return user_index, np.arange(self.n_items), self.evaluation_matrix[user_index].toarray().squeeze().astype('float32')
